feat: added swagger.json static generation
